+++ common/cryptomsg.h
@@ -1,21 +1,12 @@
 #ifndef LIGHTNING_COMMON_CRYPTOMSG_H
 #define LIGHTNING_COMMON_CRYPTOMSG_H
 #include "config.h"
-#include <bitcoin/privkey.h>
 #include <ccan/short_types/short_types.h>
 #include <ccan/tal/tal.h>
+#include <common/crypto_state.h>
 
 struct io_conn;
 struct peer;
-
-struct crypto_state {
-	/* Received and sent nonces. */
-	u64 rn, sn;
-	/* Sending and receiving keys. */
-	struct secret sk, rk;
-	/* Chaining key for re-keying */
-	struct secret s_ck, r_ck;
-};
 
 struct peer_crypto_state {
 	struct crypto_state cs;
@@ -45,9 +36,6 @@ struct io_plan *peer_write_message(struct io_conn *conn,
 				   const u8 *msg,
 				   struct io_plan *(*next)(struct io_conn *,
 							   struct peer *));
-
-void towire_crypto_state(u8 **pptr, const struct crypto_state *cs);
-void fromwire_crypto_state(const u8 **ptr, size_t *max, struct crypto_state *cs);
 
 /* Low-level functions for sync comms: doesn't discard unknowns! */
 u8 *cryptomsg_encrypt_msg(const tal_t *ctx,
